+++ sharktank/tests/layers/paged_llama_attention_block_test.py
@@ -6,7 +6,7 @@
 
 import pytest
 from sharktank.layers.configs.llm_configs import LlamaHParams, LlamaModelConfig
-from sharktank.layers.attention import CacheAllocation
+from sharktank.layers.kv_cache import CacheAllocation
 from sharktank.layers.paged_attention import build_cache
 import unittest
 import torch
